created new component for event list. minor tweaks
