+++ webapp/src/app/projects/view-project/event-list/event-list.component.html
@@ -1,0 +1,17 @@
+<div>
+    <h3>Development log <svg (click)="changeSort()" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-arrow-up-down-icon lucide-arrow-up-down"><path d="m21 16-4 4-4-4"/><path d="M17 20V4"/><path d="m3 8 4-4 4 4"/><path d="M7 4v16"/></svg>
+    </h3>
+</div>
+<div class="event-list">
+    <app-add-event [editable]="editable()" (create)="onAddEvent($event)" ></app-add-event>
+
+    @for (event of sortedEvents(); track event.id) {
+        <app-view-project-event
+            [canDelete]="editable()"
+            [event]="event"
+            [editable]="editable()"
+            (delete)="onDeleteEvent(event)"
+            (value)="updateEventDesc($event, event)"
+            (visited)="onVisited(event)"/>
+    }
+</div>
